only delete charges on delink if they exist

+++ app/models/linked_account.rb
@@ -33,8 +33,10 @@ class LinkedAccount < ActiveRecord::Base
   def delink
     #AccountDelinker.perform_async(id)
     update( destroyed_at: Time.now )
-    charges.recurring.update_all(linked_account_id: nil)
-    charges.delete_all
+    if charges.present?
+      charges.recurring.update_all(linked_account_id: nil)
+      charges.delete_all
+    end
   end
 
   def account_name
